pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "float-stream-avatar"
dynamic = ["version"]
description = "Float Stream Avatar"
requires-python = ">=3.10,<3.11"
dependencies = [
    "pyyaml",
    "opencv-python",
    "pandas",
    "tqdm",
    "matplotlib",
    "flow-vis",
    "librosa",
    "transformers==4.30.2",
    "albumentations==1.4.15",
    "albucore==0.0.16",
    "torchdiffeq==0.2.5",
    "timm==1.0.9",
    "face_alignment==1.4.1",
    "av==12.0.0",
    "numpy<2",
    "livekit-agents~=1.0",
    "moviepy==2.2.1",
    "uvicorn",
    "fastapi",
    "python-dotenv",
    "python-multipart>=0.0.20",
]

[tool.setuptools]
packages = ["models"]
